FIX ReadOnlyRestCollection inheritance with RestCollection
A RestCollection (Read/Write) is a ReadOnlyRestCollection (Read), and so forth for the interfaces and the web controllers
