+++ Domain/RDD.Domain/RDD.Domain.csproj
@@ -2,8 +2,9 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <Version>2.0.0</Version>
   </PropertyGroup>
-
+  
   <ItemGroup>
     <PackageReference Include="LinqKit.Microsoft.EntityFrameworkCore" Version="1.1.10" />
     <PackageReference Include="Microsoft.Extensions.Primitives" Version="2.0.0" />
