+++ Web/RDD.Web/RDD.Web.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <Version>2.0.0</Version>
   </PropertyGroup>
 
   <ItemGroup>
